+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+### Fixed
+- thresholds of other user in shared report
+
+
 ## 4.8.0 - 2023-03-16
 ### Added
 - column picker for data sources #337